Rust Iterator/Combinator Helpers
This is a library for iterator and combinator chain helpers that go
beyond the set provided by the stdlib or itertools. It includes
more abstract ideas like splitting off a filtered iterator that
excludes divergent values (like Err or None) but re-including them
after, or using arbitrary `fold`-style operations on the same sort
of sequence without needing to `collect()` into a temporary vector.